feat : [#9] Member 와 MemberAddress 분리
