+++ src/main/java/com/example/makedelivery/domain/member/repository/MemberRepository.java
@@ -1,6 +1,7 @@
 package com.example.makedelivery.domain.member.repository;
 
 import com.example.makedelivery.domain.member.model.entity.Member;
+import com.example.makedelivery.domain.member.model.entity.Member.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
