+++ src/main/java/com/example/makedelivery/domain/member/service/MemberService.java
@@ -27,7 +27,6 @@ import static com.example.makedelivery.common.exception.ExceptionEnum.*;
 public class MemberService {
 
     private final MemberRepository memberRepository;
-    private final MemberAddressRepository memberAddressRepository;
     private final PasswordEncoder passwordEncoder;
 
     @Transactional
@@ -93,52 +92,6 @@ public class MemberService {
     @Transactional
     public void deleteMember(Member member) {
         member.deleteMemberStatus(LocalDateTime.now());
-    }
-
-    @Transactional
-    public void addAddress(Member member, MemberAddressRequest request) {
-        MemberAddress address = MemberAddressRequest.toEntity(request, member.getId());
-        Long addressId = memberAddressRepository.save(address).getId();
-        // 아직 메인 주소가 등록되어있지 않다면, 자동으로 메인 주소를 등록해줍니다.
-        if (member.getMainAddressId() == null) member.updateMainAddress(addressId, LocalDateTime.now());
-    }
-
-    @Transactional
-    public void updateMemberAddress(Member member, Long addressId) {
-        member.updateMainAddress(addressId, LocalDateTime.now());
-    }
-
-    @Transactional
-    public void deleteAddress(Member member, Long addressId) {
-
-        if (Objects.equals(member.getMainAddressId(), addressId)) throw new ApiException(MAIN_ADDR_DELETE); // 메인 주소는 삭제할 수 없습니다.
-
-        MemberAddress memberAddress = memberAddressRepository
-                .findMemberAddressesByIdAndMemberId(addressId, member.getId())
-                .orElseThrow(() -> new ApiException(ADDR_NOT_FOUND));
-
-        memberAddress.deleteAddress(LocalDateTime.now());
-    }
-
-    @Transactional(readOnly = true)
-    public List<MemberAddressResponse> getMyAddressList(Member member) {
-        // Status 가 DEFAULT 인 주소만 불러옵니다.
-        return memberAddressRepository
-                .findAllByMemberIdAndStatus(member.getId(), Status.DEFAULT)
-                .orElse(List.of())
-                .stream()
-                .map(MemberAddressResponse::toMemberAddressResponse)
-                .toList();
-    }
-
-    /**
-     * 처음 주소 삭제시에는 Status 가 Deleted 상태가 됩니다. ( 현재 배달 중 및 서비스 로직의 문제가 생기지 않게 하기 위해서 )
-     * 이 후 스케쥴러로 Status 가 Deleted 상태인 주소들은 DB 에서 삭제합니다.
-     * 해당 서비스 로직은 스케쥴러에서 호출합니다.
-     */
-    @Transactional
-    public void deleteAllAddressDeleteStatus() {
-        memberAddressRepository.deleteAllByStatus(Status.DELETED);
     }
 
 }
